Move RuleElementSource to own file
This makes it easier to generate schemas with and makes possible
circular dependencies more clear.
