+++ tsconfig-ruleElementSource.json
@@ -1,0 +1,10 @@
+{
+  "compilerOptions": {
+    "module": "ES2020",
+    "moduleResolution": "Node",
+    "noEmit": true,
+    "strict": true,
+    "target": "ES2020"
+  },
+  "files": ["./src/typescript/ruleEngine/ruleElementSource.d.ts"]
+}
